requirements.txt
allure-pytest==2.9.45
allure-python-commons==2.9.45
attrs==21.4.0
certifi==2021.10.8
charset-normalizer==2.0.10
distlib==0.3.4
enum34==1.1.10
filelock==3.4.2
idna==3.3
importlib-metadata==4.10.0
iniconfig==1.1.1
lxml==4.7.1
namedlist==1.8
packaging==21.3
platformdirs==2.4.1
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.7
pytest==6.2.5
pytest-html==3.1.1
pytest-metadata==1.11.0
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
toml==0.10.2
typing_extensions==4.0.1
urllib3==1.26.8
virtualenv==20.13.0
zipp==3.7.0